open online help page for CLI commands
